+++ examples/lambda-container-image/index.ts
@@ -17,9 +17,9 @@ const role = new aws.iam.Role("demo-role", {
     })
 }, providerOpts);
 
-const lambdaFullAccess = new aws.iam.RolePolicyAttachment("lambdaFullAccess", {
-    role: role.name,
-    policyArn: aws.iam.ManagedPolicies.AWSLambdaFullAccess
+const lambdaFullAccessCopyAttachment = new aws.iam.RolePolicyAttachment("lambdaFullAccessCopyAttachment", {
+    role: role,
+    policyArn: aws.iam.ManagedPolicy.LambdaFullAccess,
 }, providerOpts)
 const ecsFullAccess = new aws.iam.RolePolicyAttachment("ecsFullAccess", {
     role: role.name,