+++ CHANGELOG.md
@@ -5,8 +5,12 @@ CHANGELOG
 * Upgrade to Pulumi v2.18.0
   * This introduces package/module registration for Go SDK.
 * Update auto-naming rules for RDS Cluster `clusterIdentifier` to follow specifications.
-* Add support for `AWSLambda_FullAccess` as a replacement for `AWSLambdaFullAccess` IAM policy.
-* Add support for `AWSLambda_ReadOnlyAccess` as a replacement for `AWSLambdaReadOnlyAccess` IAM policy.
+* Add support for `LambdaFullAccess` as a replacement for `AWSLambdaFullAccess` IAM policy.
+* Add support for `LambdaReadOnlyAccess` as a replacement for `AWSLambdaReadOnlyAccess` IAM policy.
+* Update `aws.lambda.CallbackFunction` to accept a new parameter `includeDeprecatedLambdaFullAccessPolicy`. If this parameter
+  is set or if no iam.Policies are set to the CallbackFunction then we will create an IAM Policy that replicates
+  the now deprecated `AWSLambdaFullAccess` policy. This is needed to allow `aws.sqs.QueueEventSubscription` and 
+  `aws.kinesis.StreamEventSubscription` continue to work
 
 ---